+++ packages/renderer/src/lib/ui/Link.svelte
@@ -2,7 +2,7 @@
 import { createEventDispatcher, onMount } from 'svelte';
 import Fa from 'svelte-fa';
 import { router } from 'tinro';
-import { isFontAwesomeIcon } from '/@/lib/ui/iconUtils';
+import { isFontAwesomeIcon } from '/@/lib/ui/icon-utils';
 
 export let internalRef: string | undefined = undefined;
 export let externalRef: string | undefined = undefined;
